+++ R/RFT_vertex_analysis.R
@@ -19,10 +19,11 @@
 #' @param contrast A N x 1 numeric vector or object containing the values of the predictor of interest. Its length should equal the number of subjects in model (and can be a single column from model). The cluster-thresholded t-stat maps will be estimated only for this predictor. 
 #' @param random A N x 1 numeric vector or object containing the values of the random variable (optional). Its length should be equal to the number of subjects in model (it should NOT be inside the model data.frame).
 #' @param formula An optional string or formula object describing the predictors to be fitted against the surface data, replacing the model, contrast, or random arguments. If this argument is used, the formula_dataset argument must also be provided.
-#' - The dependent variable is not needed, as it will always be the surface data values. 
+#' - The dependent variable (DV) is not needed, and the formula will start with ~. The DV will be the surface data value by default, but it can be swapped with contrast as IV via the "inverse" argument.
 #' - The first independent variable in the formula will always be interpreted as the contrast of interest for which to estimate cluster-thresholded t-stat maps. 
 #' - Only one random regressor can be given and must be indicated as '(1|variable_name)'.
 #' @param formula_dataset An optional data.frame object containing the independent variables to be used with the formula (the IV names in the formula must match their column names in the dataset).
+#' @param inverse A boolean object stating whether to set the surface data as predictor of the contrast variable, instead of as dependent variable (default is FALSE). Other covariates in the model remain independent variables. This makes modelling slower.
 #' @param surf_data A N x V matrix object containing the surface data (N row for each subject, V for each vertex), in fsaverage5 (20484 vertices), fsaverage6 (81924 vertices), fslr32k (64984 vertices) or hippocampal (14524 vertices) space. See also Hipvextract(), SURFvextract() or FSLRvextract output formats. Alternatively, a string object containing the path to the surface object (.rds file) outputted by extraction functions may be given.
 #' @param p A numeric object specifying the p-value to threshold the results (Default is 0.05)
 #' @param atlas A numeric integer object corresponding to the atlas of interest.  1=Desikan, 2=Destrieux-148, 3=Glasser-360, 4=Schaefer-100, 5=Schaefer-200, 6=Schaefer-400. Set to `1` by default. This argument is ignored for hippocampal surfaces.
@@ -57,7 +58,7 @@
 #' @export
 
 ##vertex wise analysis with mixed effects
-RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, surf_data, p=0.05, atlas=1, smooth_FWHM, VWR_check=TRUE)  ## atlas: 1=Desikan, 2=Schaefer-100, 3=Schaefer-200, 4=Glasser-360, 5=Destrieux-148; ignored for hippocampal surfaces
+RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, inverse=FALSE, surf_data, p=0.05, atlas=1, smooth_FWHM, VWR_check=TRUE)  ## atlas: 1=Desikan, 2=Schaefer-100, 3=Schaefer-200, 4=Glasser-360, 5=Destrieux-148; ignored for hippocampal surfaces
 {
   #gets surface matrix if is surf_data is a list or path
   surf_data=get_surf_obj(surf_data)
@@ -126,8 +127,9 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   . <- MixedEffect <- NULL
   
   ##import python libaries
+  #version of SLM that allows to specify the directory for the
+  #fetch_template_surface option
   reticulate::source_python(paste0(system.file(package='VertexWiseR'),'/python/brainstat.stats.SLM_VWR.py'))
-  
 
   ##fitting model
   #preparing mask for model
@@ -148,13 +150,8 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   data_dir=paste0(brainstat_data_path,'/brainstat_data/surface_data/')
   
   #define model to fit
-  if(missing(random)) {model0=FixedEffect(model, "_check_categorical" = FALSE)}
+  if(is.null(random)) {model0=FixedEffect(model, "_check_categorical" = FALSE)}
   else {model0=MixedEffect(ran = as.factor(random),fix = model,"_check_categorical" = FALSE)}
-  
-
-  #read version of SLM that allows to specify the directory for the
-  #fetch_template_surface option
-  
   
   model=SLM(model = model0,
             contrast=contrast,
@@ -162,10 +159,8 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
             mask=mask,
             correction=c("fdr", "rft"),
             cluster_threshold=p,
-            data_dir=data_dir)
-  
-  #fit will fetch parcellation data in a different place
-  model$data_dir=paste0(brainstat_data_path,'/brainstat_data/parcellation_data/')
+            data_dir=data_dir,
+            inverse=inverse)
   
   #fit model
   SLM$fit(model,surf_data)
@@ -175,7 +170,9 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   
   ##extracting positive results
   cluster_pos=reticulate::py_to_r(model$P[["clus"]][[1]]) #pulling out results from brainstat's output
-  cluster_pos=cluster_pos[cluster_pos$P<p,] #removing clusters that are not significant
+  cluster_pos=cluster_pos[which(cluster_pos$P<p & !is.na(cluster_pos$P)),] #removing clusters that are not significant
+  if(NROW(cluster_pos)!=0)
+  {cluster_pos$clusid=1:nrow(cluster_pos)} #safe renumbering as NaN Ps can be introduced first in the list
   
   #extracting positive cluster map
   pos_clusterIDmap=model$P$clusid[[1]]
@@ -187,6 +184,11 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   } else
   {
     #creating new result variables in the cluster_pos objects
+    
+    #convert elements to numeric instead of list
+    if(inherits(cluster_pos$P,"list")){cluster_pos$P=unlist(cluster_pos$P)}
+    if (inherits(cluster_pos$clusid,'list')){cluster_pos$clusid=unlist(cluster_pos$clusid)}
+    
     cluster_pos$P=round(cluster_pos$P,3)
     cluster_pos$P[cluster_pos$P==0]="<0.001"
     cluster_pos=cluster_pos[ , !(names(cluster_pos) %in% "resels")] #removing the 'resels' column from the original brainstat output
@@ -217,7 +219,9 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   
   ##extracting negative results
   cluster_neg=reticulate::py_to_r(model$P[["clus"]][[2]]) #pulling out results from brainstat's output
-  cluster_neg=cluster_neg[cluster_neg$P<p,] #removing clusters that are not significant
+  cluster_neg=cluster_neg[which(cluster_neg$P<p & !is.na(cluster_neg$P)),] #removing clusters that are not significant
+  if(NROW(cluster_neg)!=0)
+  {cluster_neg$clusid=1:nrow(cluster_neg)} #safe renumbering as NaN Ps can be introduced first in the list
   
   #extracting negative cluster map
   neg_clusterIDmap=model$P$clusid[[2]]
@@ -227,6 +231,11 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
     neg_clusterIDmap=rep(0, NCOL(surf_data))
   } else
   { #creating new result variables in the cluster_pos objects
+    
+    #convert elements to numeric instead of list
+    if(inherits(cluster_neg$P,"list")){cluster_neg$P=unlist(cluster_neg$P)}
+    if (inherits(cluster_neg$clusid,'list')){cluster_neg$clusid=unlist(cluster_neg$clusid)}
+    
     cluster_neg$P=round(cluster_neg$P,3)
     cluster_neg$P[cluster_neg$P==0]="<0.001"
     cluster_neg=cluster_neg[ , !(names(cluster_neg) %in% "resels")] #removing the 'resels' column from the original brainstat output