+++ R/plot_surf3d.R
@@ -4,12 +4,14 @@
 #'
 #' @param surf_data  A numeric vector (length of V) 
 #' @param surf_color  color of the cortical surface. Set to `'grey'` by default
-#' @param cmap A string vector containing 2 to 4 color names/codes specifying the colors to be used for the color scale. See `RColorBrewer::display.brewer.all()` for all possible cmap options. If none are specified, appropriate colors will be automatically selected according to `range(surf_data)`
+#' @param cmap A string vector containing 2 to 4 color names/codes specifying the colors to be used for the color scale; or a single string object with the name of a color map listed in `RColorBrewer::display.brewer.all()`. If none are specified, appropriate colors will be automatically selected according to `range(surf_data)`
 #' @param limits A combined pair of numeric vector composed of the lower and upper color scale limits of the plot. When left unspecified, the symmetrical limits `c(-max(abs(surf_dat),max(abs(surf_dat)))` will be used. 
 #' @param atlas atlas used for identifying region labels. 1=Desikan, 2=Destrieux-148, 3=Glasser-360, 4=Schaefer-100, 5=Schaefer-200, 6=Schaefer-400. Set to `1` by default. This argument is ignored for hippocampal surfaces.
 #' @param hemi A string specifying the hemisphere to plot. Possible values are `l` (left), `r` (right) or `b` (both).
 #' @param medial_gap A numeric value specifying the amount of gap (in MNI coordinate units) to separate the left and right hemispheres. Set to `0` (no gap between hemispheres) by default. In order to view the medial surfaces clearly, it is recommended that this value is set to `20`. This argument is ignored if `hemi!='b'`
 #' @param orientation_labels A boolean object specifying if orientation labels are to be displayed. Set to `TRUE` by default
+#' @param plot_grid A boolean object specifying whether to plot the orientation grid or not (default is `TRUE`).
+#' @param transparent_bg A boolean object specifying whether to get a transparent background upon saving the image (default is `FALSE`, white background).
 #' @param VWR_check A boolean object specifying whether to check and validate system requirements. Default is TRUE.
 #'
 #' @returns a plot_ly object
@@ -17,10 +19,12 @@
 #' surf_data = runif(20484);
 #' plot_surf3d(surf_data = surf_data, VWR_check=FALSE)
 #' @importFrom plotly plot_ly add_trace layout
+#' @importFrom grDevices colorRampPalette
+#' @importFrom RColorBrewer brewer.pal.info brewer.pal
 #' @export
 ######################################################################################################################################################
 ######################################################################################################################################################
-plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b",medial_gap=0,orientation_labels=TRUE,VWR_check=TRUE)
+plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b",medial_gap=0,orientation_labels=TRUE,VWR_check=TRUE,plot_grid=TRUE,transparent_bg=FALSE)
 {
   #Check required python dependencies. If files missing:
   #Will prompt the user to get them in interactive session 
@@ -43,6 +47,21 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
       if(range(surf_data,na.rm = TRUE)[1]>=0)  {cmap=c("#A51122","#F5FACD")}
       else if (range(surf_data,na.rm = TRUE)[2]<=0)  {cmap=c("#324DA0","#E7F1D5")}
       else  {cmap=c("#E7F1D5","#324DA0","#A51122","#F5FACD")}  
+    }
+    #build RColorBrewer colormaps manually to make sure plotly renders them properly on the mesh3d
+    if (length(cmap)==1)
+    {  
+      if (cmap %in% rownames(RColorBrewer::brewer.pal.info))
+      {#extract hex palette from selected cmap
+        pal <- grDevices::colorRampPalette(RColorBrewer::brewer.pal(RColorBrewer::brewer.pal.info[cmap, "maxcolors"], cmap))
+        cols <- pal(100) #split color vector into palette of 100
+        #Plotly expects a colorscale as a list of pairs: [fraction, color]. So for each color, get fraction between 0 and 1 and its matching hex code
+        cmap <- lapply(seq_along(cols), function(i) 
+        {
+          list((i-1)/(length(cols)-1), #fraction
+               cols[i]) #hex
+        })
+      }
     }
     # enabling custom color scales
     if(length(cmap)==2) {cmap=list(list(0,cmap[1]), list(1,cmap[2]))} 
@@ -92,7 +111,7 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
     maxlimit=max(abs(range(face.stat,na.rm = TRUE)))
     if(missing(limits)) 
     {
-      limits.range=range(face.stat,na.rm = T)
+      limits.range=range(face.stat,na.rm = TRUE)
       if(limits.range[1]>=0) {limits=c(0,limits.range[2])} ##if image contains all positive values
       else if(limits.range[2]<=0) {limits=c(limits.range[1],0)} ##if image contains all negative values
       else if(limits.range[1]<0 & limits.range[2]>0){limits=c(-maxlimit,maxlimit)} ##symmetrical limits will be used if image contains both positive and negative values
@@ -147,16 +166,18 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
                         z = coords[,3],
                         i = tri[, 1] - 1,  # plotly uses 0-based indexing, so subtract 1
                         j = tri[, 2] - 1,
-                        k = tri[, 3] - 1,facecolor=rep(surf_color,NROW(tri)))
+                        k = tri[, 3] - 1,
+                        facecolor=rep(surf_color,NROW(tri)))
   
   ##overlay statistical map on cortical surface
-    fig=add_trace(fig,type = 'mesh3d',
+    fig=plotly::add_trace(fig,type = 'mesh3d',
                   i = tri[face.stat.non0.idx,][, 1] - 1,  # plotly uses 0-based indexing, so subtract 1
                   j = tri[face.stat.non0.idx,][, 2] - 1,
                   k = tri[face.stat.non0.idx,][, 3] - 1,
-                  intensitymode="cell",intensity=face.stat[face.stat.non0.idx],
+                  intensitymode="cell",
+                  intensity=face.stat[face.stat.non0.idx],
                   colorscale = cmap,
-                  cauto = F,
+                  cauto = FALSE,
                   cmin = limits[1],
                   cmax = limits[2])
 
@@ -169,7 +190,7 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
     customdata[(mid.idx+1):n_vert]=customdata[(mid.idx+1):n_vert]-medial_gap
     
     ##add mouse-over text
-    fig=add_trace(fig,text=ROI.text,hovertext=surf_data,intensitymode="vertex", intensity=0, opacity=0,showscale= F,
+    fig=add_trace(fig,text=ROI.text,hovertext=surf_data,intensitymode="vertex", intensity=0, opacity=0,showscale=FALSE,
                   x = coords[,1],
                   y = coords[,2],
                   z = coords[,3],
@@ -182,7 +203,7 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
                                       "statistic:%{hovertext:.2f}<extra></extra>"))
   } else
   {
-    fig=add_trace(fig,text=ROI.text,intensitymode="vertex", intensity=0, opacity=0,showscale= F,
+    fig=add_trace(fig,text=ROI.text,intensitymode="vertex", intensity=0, opacity=0,showscale= FALSE,
                   x = coords[,1],
                   y = coords[,2],
                   z = coords[,3],
@@ -194,22 +215,47 @@ plot_surf3d=function(surf_data, surf_color="grey",cmap,limits, atlas=1, hemi="b"
                                       "MNI coords: %{x:.1f},%{y:.1f},%{z:.1f}<br>",
                                       "statistic:%{customdata:.2f}<extra></extra>"))
   }
-  ##axis parameters
-  fig=plotly::layout(fig,
-                   hoverlabel = list(align = "left"),
-                   scene = list(camera=list(eye = list(x = 0, y = 1.5, z = 1.5)),
-                                xaxis = list(showgrid = T,showticklabels=T,showspikes=F,zeroline=F, title=""),
-                                yaxis = list(showgrid = T,showticklabels=T,showspikes=F,zeroline=F, title=""),
-                                zaxis = list(showgrid = T,showticklabels=T,showspikes=F,zeroline=F, title="")))
-  
-  ##add optional orientation labels
-  if(orientation_labels==T)
-  {
-    axx = list(ticketmode = 'array',ticktext = xlab,tickvals = range(coords[,1]))
-    axy = list(ticketmode = 'array',ticktext = c("Posterior","Anterior"),tickvals = range(coords[,2]))
-    axz = list(ticketmode = 'array',ticktext = c("Inferior","Superior"),tickvals = range(coords[,3]))
     
-    fig = layout(fig,scene = list(xaxis=axx,yaxis=axy,zaxis=axz))
+    
+  if (plot_grid==TRUE)
+  {
+    ##axis parameters
+    fig=plotly::layout(fig,
+                     hoverlabel = list(align = "left"),
+                     scene = list(camera=list(eye = list(x = 0, y = 1.5, z = 1.5)),
+                                  xaxis = list(showgrid = TRUE,showticklabels=TRUE,showspikes=FALSE,zeroline=FALSE, title=""),
+                                  yaxis = list(showgrid = TRUE,showticklabels=TRUE,showspikes=FALSE,zeroline=FALSE, title=""),
+                                  zaxis = list(showgrid = TRUE,showticklabels=TRUE,showspikes=FALSE,zeroline=FALSE, title="")))
+    
+    ##add optional orientation labels
+    if(orientation_labels==TRUE)
+    {
+      axx = list(ticketmode = 'array',ticktext = xlab,tickvals = range(coords[,1]))
+      axy = list(ticketmode = 'array',ticktext = c("Posterior","Anterior"),tickvals = range(coords[,2]))
+      axz = list(ticketmode = 'array',ticktext = c("Inferior","Superior"),tickvals = range(coords[,3]))
+      
+      fig = layout(fig,scene = list(xaxis=axx,yaxis=axy,zaxis=axz))
+    }
+  }
+  else #to remove grid
+  {  fig <- plotly::layout(
+      fig,
+      scene = list(
+        xaxis = list(showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE, visible = FALSE),
+        yaxis = list(showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE, visible = FALSE),
+        zaxis = list(showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE, visible = FALSE),
+        bgcolor = "rgba(0,0,0,0)")
+  )
+  }
+    
+  #to get transparent background
+  if (transparent_bg==TRUE)
+  { 
+    fig <- plotly::layout(
+    fig,
+    paper_bgcolor = "rgba(0,0,0,0)",
+    plot_bgcolor  = "rgba(0,0,0,0)" 
+    )
   }
   return(fig)
 }